trello-test-backend/requirements.txt
Flask
graphene
Flask-GraphQL
boto3
python-dotenv
# Include other dependencies as needed
